% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnaseq_datasets.R
\docType{data}
\name{rnaseq}
\alias{ACC.rnaseq}
\alias{ACC.rnaseq,BLCA.rnaseq,BRCA.rnaseq,CESC.rnaseq,CHOL.rnaseq,COAD.rnaseq,COADREAD.rnaseq,DLBC.rnaseq,ESCA.rnaseq,FPPP.rnaseq,GBM.rnaseq,GBMLGG.rnaseq,HNSC.rnaseq,KICH.rnaseq,KIPAN.rnaseq,KIRC.rnaseq,KIRP.rnaseq,LAML.rnaseq,LGG.rnaseq,LIHC.rnaseq,LUAD.rnaseq,LUSC.rnaseq,MESO.rnaseq,OV.rnaseq,PAAD.rnaseq,PCPG.rnaseq,READ.rnaseq,SKCM.rnaseq,STAD.rnaseq,STES.rnaseq,TGCT.rnaseq,THCA.rnaseq,THYM.rnaseq,UCEC.rnaseq,UCS.rnaseq,UVM.rnaseq}
\alias{BLCA.rnaseq}
\alias{BRCA.rnaseq}
\alias{CESC.rnaseq}
\alias{CHOL.rnaseq}
\alias{COAD.rnaseq}
\alias{COADREAD.rnaseq}
\alias{DLBC.rnaseq}
\alias{ESCA.rnaseq}
\alias{FPPP.rnaseq}
\alias{GBM.rnaseq}
\alias{GBMLGG.rnaseq}
\alias{HNSC.rnaseq}
\alias{KICH.rnaseq}
\alias{KIPAN.rnaseq}
\alias{KIRC.rnaseq}
\alias{KIRP.rnaseq}
\alias{LAML.rnaseq}
\alias{LGG.rnaseq}
\alias{LIHC.rnaseq}
\alias{LUAD.rnaseq}
\alias{LUSC.rnaseq}
\alias{MESO.rnaseq}
\alias{OV.rnaseq}
\alias{PAAD.rnaseq}
\alias{PCPG.rnaseq}
\alias{READ.rnaseq}
\alias{SKCM.rnaseq}
\alias{STAD.rnaseq}
\alias{STES.rnaseq}
\alias{TGCT.rnaseq}
\alias{THCA.rnaseq}
\alias{THYM.rnaseq}
\alias{UCEC.rnaseq}
\alias{UCS.rnaseq}
\alias{UVM.rnaseq}
\alias{rnaseq}
\title{rnaseq datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.rnaseq

BRCA.rnaseq

BLCA.rnaseq

BRCA.rnaseq

CESC.rnaseq

CHOL.rnaseq

COAD.rnaseq

COADREAD.rnaseq

DLBC.rnaseq

ESCA.rnaseq

FPPP.rnaseq

GBM.rnaseq

GBMLGG.rnaseq

HNSC.rnaseq

KICH.rnaseq

KIPAN.rnaseq

KIRC.rnaseq

KIRP.rnaseq

LAML.rnaseq

LGG.rnaseq

LIHC.rnaseq

LUAD.rnaseq

LUSC.rnaseq

MESO.rnaseq

OV.rnaseq

PAAD.rnaseq

PCPG.rnaseq

READ.rnaseq

SKCM.rnaseq

STAD.rnaseq

STES.rnaseq

TGCT.rnaseq

THCA.rnaseq

THYM.rnaseq

UCEC.rnaseq

UCS.rnaseq

UVM.rnaseq
}
\description{
Package provides rnaseq datasets from The Cancer Genome
Atlas Project for all cohorts types from \url{http://
gdac.broadinstitute.org/}. Data were downloaded using
\link{RTCGA-package}. The process is described here:
\url{http://mi2-warsaw.github.io/RTCGA.data/}.
}
\details{
rnaseq data format is explained here \url{https://wiki.nci.nih.gov/display/TCGA/RNASeq+Version+2}.
}
\keyword{datasets}

